src/raspberry/host_redis_mirror/requirements.txt
redis==3.5.3
pyyaml==5.4.1
